requirements.txt
PyQt5~=5.14.1
pyzbar~=0.1.8
Pillow~=7.0.0